+++ TP Especial/tpmonopatines/seguimiento/src/main/java/main/paradas/servicios/paradaService.java
@@ -1,10 +1,12 @@
 package main.paradas.servicios;
 
+import main.paradas.dtos.DtoParada;
 import main.paradas.model.parada;
 import main.paradas.repository.paradaRepository;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,22 +20,42 @@ public class paradaService {
     private paradaRepository paradaRepository;
 
     @Transactional
-    public parada save(parada parada) {
-        return paradaRepository.save(parada);
+    public DtoParada save(parada parada) {
+        parada resultado = paradaRepository.save(parada);
+        return new DtoParada(resultado.getNro_parada(), resultado.getNombre(), resultado.getParada_permitida());
     }
 
     @Transactional
-    public List<parada> findAll() {
-        return paradaRepository.findAll();
+    public List<DtoParada> findAll() throws Exception {
+        var resultados =  paradaRepository.findAll();
+        try {
+            return resultados.stream().map(resultado -> new DtoParada(resultado.getNro_parada(),resultado.getNombre(),resultado.getParada_permitida())).collect(Collectors.toList());
+        } catch (Exception e) {
+            throw new Exception(e.getMessage());
+        }
     }
 
     @Transactional
-    public Optional<parada> findById(Long id) {
-        return paradaRepository.findById(id);
+    public DtoParada findById(Long id) {
+        Optional<parada> res = paradaRepository.findById(id);
+        if (res.isPresent()) {
+            parada resultado = res.get();
+            return new DtoParada(resultado.getNro_parada(), resultado.getNombre(), resultado.getParada_permitida());
+        } else {
+            return null;
+        }
+
     }
 
     @Transactional
-    public void deleteById(Long id) {
-        paradaRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        Optional<parada> admin = paradaRepository.findById(id);
+        if (admin.isPresent()) {
+            paradaRepository.deleteById(id);
+            return true;
+        }
+        else{
+            return false;
+        }
     }
 }
